feat: config setup, overlay relocation, dynamic language load, and caption author support
