+++ .gitignore
@@ -11,3 +11,4 @@ __pycache__
 node_modules/
 frontend/build/
 dist/
+config.js
