+++ chrome-extension-overlay/background/config.js.example
@@ -1,0 +1,6 @@
+// config.js (DO NOT commit this to version control)
+export const CONFIG = {
+  AZURE_TRANSLATOR_KEY: "",
+  AZURE_TRANSLATOR_REGION: "",
+  AZURE_TRANSLATOR_ENDPOINT: "https://api.cognitive.microsofttranslator.com"
+};
