+++ chrome-extension-overlay/popup/popup.html
@@ -37,7 +37,7 @@
     
     <!-- Settings Section -->
     <div class="settings-section">
-      <div class="setting-group">
+      <!-- <div class="setting-group">
         <label for="sourceLanguage">Source Language:</label>
         <select id="sourceLanguage" class="select-input">
           <option value="en-US">English (US)</option>
@@ -54,7 +54,7 @@
           <option value="ar-SA">Arabic</option>
           <option value="hi-IN">Hindi</option>
         </select>
-      </div>
+      </div> -->
       
       <div class="setting-group">
         <label for="targetLanguage">Translate to:</label>
@@ -77,7 +77,7 @@
       
       <div class="setting-group">
         <label class="checkbox-label">
-          <input type="checkbox" id="showOriginal" checked>
+          <input type="checkbox" id="showOriginal">
           <span class="checkmark"></span>
           Show original text
         </label>
